edit.c typecast param nudging

+++ src/ui/edit.c
@@ -38,7 +38,7 @@ void edit_draw(EDIT *edit, int x, int y, int width, int height) {
         return;
     }
 
-    if (settings.window_baseline && y > settings.window_baseline - font_small_lineheight - SCALE(8)) {
+    if (settings.window_baseline && y > (int)settings.window_baseline - font_small_lineheight - SCALE(8)) {
         y = settings.window_baseline - font_small_lineheight - SCALE(8);
     }
 
@@ -126,7 +126,7 @@ void edit_draw(EDIT *edit, int x, int y, int width, int height) {
 }
 
 bool edit_mmove(EDIT *edit, int px, int py, int width, int height, int x, int y, int dx, int dy) {
-    if (settings.window_baseline && py > settings.window_baseline - font_small_lineheight - SCALE(8)) {
+    if (settings.window_baseline && py > (int)settings.window_baseline - font_small_lineheight - SCALE(8)) {
         y += py - (settings.window_baseline - font_small_lineheight - SCALE(8));
         py = settings.window_baseline - font_small_lineheight - SCALE(8);
     }
